Add Kingsley, Dylan, Audric to CODEOWNERS (#769)
**Kingsley**: in the last quarter, I've had to admin bypass merge requirements a few times times to unblock Florian on key 
fixes / changes. This is mainly for the rare case so that I don't have to fuck with settings just to get something minor in. I WILL NOT BE APPROVING ANYTHING MAJOR. 

**Dylan**: PostHog related approvals. Has repeatedly demonstrated deep understanding and competency with React patterns that we use and has implemented multiple high-quality features in the codebase.

**Audric**: Built clinical & ops portal 0 -> 1. Would be helpful for improving velocity given his deep knowledge of the codebase and React. 

I think there's an understanding with everyone here not to "LGTM" massive feature changes or anything billing/auth related. The idea here is to unblock velocity for minor tickets (throughput).